requirements.txt
python-datauri==0.2.8
